ref: Refactor setColorScheme to reuse in stories
